This should fix it...
The problem was that StringContext wasn't resolved with a `f` in it.
The thing being that the SBT's `scala.StringContext` class from the `sbt-launch.jar` doesn't indeed define it :-S.

Okay, but where is included this jar in the classpath when running `play run`? In the JavaPlatform (from scala jar), which its `result` will contain the user java path.
This path is actually, the only `sbt-launch.jar` when running `play run`.

This can be avoided by setting (a contrary from previous messages): settings.usejavacp.value to false.

Now need some cleanup and detect if the running mode must be detected or, `dist` will still work