fixes #21
This one support just some rundimentary combinors (the weak way): | and &&.
Best would be to use `sys.process` directly from a *code cell* for more elaborated shell scripts

This commit cleans a bit the sbt from unused stuffs and fixes scalaVersion to 2.10.4 everywhere.

It also allows for specifying -Dspark.version and -Dhadoop.version in order to build the notebook against specific versions.
This is only a quick and dirty (and probably ephemeral) support, cross builds would be better.