[WIP] Integration with Spark 1.2.0
For now, the spark-repl 1.2.0 is still not published... https://issues.apache.org/jira/browse/SPARK-4923
So a custom build is integrated in the project (see lib folder) that package only the required library for it to work (spark it-self will be imported by others)

This must be temporary, it's hackish power 2 but does the job.

Since Spark is using akka 2.3.4, we have the opportunity to migrate to the latest play version. Hence, to avoid to much complexity, the full project has been migrated.

That being said, the project will live in two different branches one for 1.1.x and another for 1.2.0x (for now).