Added Java 7 as a requirement in the documentation
I was trying to get this running inside of Vagrant with the default hashicorp/precise32 image (Ubuntu 12). `sudo apt-get install default-jdk` installs OpenJDK 6, which generates an error when you try to open the notebook in your browser.

I'm not sure the user really needs the JDK or if the JRE would suffice.