Adds support for long lived sessions but misses cross interactions
* _Running_ tabs shows the running kernels/sessions
* there is **only** one session per **path** →  that means that they share the kernel behind the scene, hence variables and context!
* if the same notebook has to be run several times in parallel (???) it must be duplicated first
* the msgs coming from the backend repl is sent over all connected web-sockets/client (page opened with the same path)

* the messages are received by the clients but there are ignored (nothing shows up in the UI) →  see below
* the web-socket for observables are still separated so they won't get updated with backend events

A web socket is established between the client (browser tab) and the backend server.

All these web sockets (for a given notebook) are also piped to a single backend repl/kernel.
So that when something gets executed in the repl, all tabs will be notified since messages are broadcasted to all of them.

However, currently, the UI will ignore them because they are sensitive to **message id**s.

Indeed, all messages issued by a notebook gets assigned an id, this id will be cached locally by the **kernel.js** component in the `_msgs_callbacks` object.
So when such message (and id) is created the function `set_callbacks_for_msg` is called to assign to the `msg_id` probable `callbacks`.

However, the functions that are called by the web socket listeners (`_handle_output_message` and `_handle_input_message`) are will check if callbacks have been attached to the message that comes out of the socket.
This is done in `_get_callbacks_for_msg`, if this guy doesn't return anything, then the message will be ignored and an even `received_unsolicited_message.Kernel` will be trigger with the **message** as payload.

That means that UIs are not in sync, so we need to either:
* change the `kernel.js` to be lenient regarding `msg_id`
* add a custom code (in `custom.js` for instance) that will take care of them!

In both case, transformation might occur at the same time, hence it's probable that we'll need some **Operational Transformation** support.