Added conf `manager.notebooks.override`
This new configuration takes the same form as `manager.notebooks.custom`.
The `custom` one will be only applied for new notebooks, however this new one will
be applied on any notebooks that are launched by the server. So to use with care.

This is for instance helpful for spark-notebook instance running on one single cluster
that would require specific libraries in the path. Or in order to bind to specific IP.