Use cell id t track spark job
This is using the great @vidma trick to track the match between jobs and cells, via job group.

So here, we forward the information of the job and group to the browser which makes use of it, very simply right now, just highlight the concerned cell.

Checkout the `Spark Jobs and Progress Bars.snb` that illustrates that even in the context of FAIR scheduler and delayed computations.

NOTE: this can be broken because of the observable instanciation again, it'll need [this](https://github.com/andypetrella/spark-notebook/pull/468) to fix it, meanwhile, clean all outputs, save and refresh the page.