Update Flow to be reusable across runs
Using cell's metadata to store the flow: boxes, links, customization and so on.
So rerunning the cell will fetch these and update the canvas.

Update save to react on add and remove to save the state accordingly.

Update the joint code to include the link instance in the event when a link is issued from a box

Updated the example with a reusable and rerunable spark pipe