Start the kernel in a read only mode with execution capabilities
The kernel control the channels to the backend server, hence it is required to see the evolution of the charts.

However, it wasn't started in read only mode, which was making the notebooks static, but too much.

Now, the kernel is stared but in read only mode, so that, the session controls what can be done (reject all execution and save actions).

Now, we can have active and passive notebooks