Serious refactoring to allow spark-notebook to run in viewer mode
This is separating spark from the main project if it is build if the viewer mode enabled (sbt `-Dviewer=true`).

Hence the spark module is then not included in the binary (however, spark are still there due to dependencies in the widgets).

This gets us closer to separate more the server from spark kernel with specific version for instance. But this wasn't the main goal here.

Downside, the `Repl` is now created using reflection... because the ReplCalculator cannot be extracted since it is used in Application.scala (cf separation remark above).

Leftover, the actions like create notebook and so forth are only disabled client side, this should be addressed on the server side too.
However, the most harmful thing, which is the opening the Repl isn't possible at all!